# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Skill Eval Platform API",
    "description": "Backend service for the unified evaluation question library.",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/auth/register": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Register",
        "operationId": "register_api_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login",
        "operationId": "login_api_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Logout",
        "operationId": "logout_api_auth_logout_post",
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Me",
        "operationId": "me_api_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/questions": {
      "get": {
        "tags": [
          "question-library"
        ],
        "summary": "List Library",
        "operationId": "list_library_api_questions_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/QuestionStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "scene_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Scene Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionLibraryResponse"
                }
              }
            }
          },
          "401": {
            "description": "未登录"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/questions/{question_id}": {
      "get": {
        "tags": [
          "question-library"
        ],
        "summary": "Get Question",
        "operationId": "get_question_api_questions__question_id__get",
        "parameters": [
          {
            "name": "question_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Question Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionDetailResponse"
                }
              }
            }
          },
          "401": {
            "description": "未登录"
          },
          "404": {
            "description": "题目不存在"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "question-library"
        ],
        "summary": "Delete Question",
        "operationId": "delete_question_api_questions__question_id__delete",
        "parameters": [
          {
            "name": "question_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Question Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "未登录"
          },
          "404": {
            "description": "题目不存在"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/questions/{question_id}/title": {
      "patch": {
        "tags": [
          "question-library"
        ],
        "summary": "Update Title",
        "operationId": "update_title_api_questions__question_id__title_patch",
        "parameters": [
          {
            "name": "question_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Question Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestionTitleRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionDetailResponse"
                }
              }
            }
          },
          "401": {
            "description": "未登录"
          },
          "404": {
            "description": "题目不存在"
          },
          "409": {
            "description": "内容版本陈旧"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/questions/{question_id}/save-regenerate": {
      "post": {
        "tags": [
          "question-library"
        ],
        "summary": "Save And Regenerate",
        "operationId": "save_and_regenerate_api_questions__question_id__save_regenerate_post",
        "parameters": [
          {
            "name": "question_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Question Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestionSaveRegenerateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationAcceptedResponse"
                }
              }
            }
          },
          "401": {
            "description": "未登录"
          },
          "404": {
            "description": "题目不存在"
          },
          "409": {
            "description": "内容版本陈旧"
          },
          "422": {
            "description": "材料内容无效"
          }
        }
      }
    },
    "/api/questions/{question_id}/criteria": {
      "patch": {
        "tags": [
          "question-library"
        ],
        "summary": "Patch Criteria",
        "operationId": "patch_criteria_api_questions__question_id__criteria_patch",
        "parameters": [
          {
            "name": "question_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Question Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CriteriaPatchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionDetailResponse"
                }
              }
            }
          },
          "401": {
            "description": "未登录"
          },
          "404": {
            "description": "题目不存在"
          },
          "409": {
            "description": "内容版本陈旧或生成中"
          },
          "422": {
            "description": "评分维度无效"
          }
        }
      }
    },
    "/api/questions/{question_id}/generation-retry": {
      "post": {
        "tags": [
          "question-library"
        ],
        "summary": "Retry Generation",
        "operationId": "retry_generation_api_questions__question_id__generation_retry_post",
        "parameters": [
          {
            "name": "question_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Question Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestionCommandRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationAcceptedResponse"
                }
              }
            }
          },
          "401": {
            "description": "未登录"
          },
          "404": {
            "description": "题目不存在"
          },
          "409": {
            "description": "状态或版本不允许重试"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/questions/{question_id}/publication": {
      "post": {
        "tags": [
          "question-library"
        ],
        "summary": "Publish",
        "operationId": "publish_api_questions__question_id__publication_post",
        "parameters": [
          {
            "name": "question_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Question Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestionCommandRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionDetailResponse"
                }
              }
            }
          },
          "401": {
            "description": "未登录"
          },
          "404": {
            "description": "题目不存在"
          },
          "409": {
            "description": "状态不允许发布"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CriteriaPatchRequest": {
        "properties": {
          "command_id": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Command Id"
          },
          "content_revision": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Content Revision"
          },
          "criteria": {
            "items": {
              "$ref": "#/components/schemas/CriterionIn"
            },
            "type": "array",
            "maxItems": 20,
            "minItems": 1,
            "title": "Criteria"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "command_id",
          "content_revision",
          "criteria"
        ],
        "title": "CriteriaPatchRequest"
      },
...
      "CriterionView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "criterion": {
            "type": "string",
            "title": "Criterion"
          },
          "pass_score": {
            "type": "integer",
            "title": "Pass Score"
          }
        },
        "type": "object",
        "required": [
          "id",
          "criterion",
          "pass_score"
        ],
        "title": "CriterionView"
      },
      "ErrorPayload": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorPayload"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorPayload"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoginRequest": {
        "properties": {
          "identifier": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Identifier"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "identifier",
          "password"
        ],
        "title": "LoginRequest"
      },
...
      "OperationAcceptedResponse": {
        "properties": {
          "question_id": {
            "type": "string",
            "title": "Question Id"
          },
          "status": {
            "$ref": "#/components/schemas/QuestionStatus"
          },
          "operation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Operation Id"
          }
        },
        "type": "object",
        "required": [
          "question_id",
          "status",
          "operation_id"
        ],
        "title": "OperationAcceptedResponse"
      },
      "QuestionCommandRequest": {
        "properties": {
          "command_id": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Command Id"
          },
          "content_revision": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Content Revision"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "command_id",
          "content_revision"
        ],
        "title": "QuestionCommandRequest"
      },
      "QuestionDetailResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "scene_id": {
            "type": "string",
            "title": "Scene Id"
          },
          "scene_name": {
            "type": "string",
            "title": "Scene Name"
          },
          "client_case_id": {
            "type": "string",
            "title": "Client Case Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "task_prompt": {
            "type": "string",
            "title": "Task Prompt"
          },
          "reference_examples": {
            "items": {
              "$ref": "#/components/schemas/ReferenceExampleView"
            },
            "type": "array",
            "title": "Reference Examples"
          },
          "bad_cases": {
            "items": {
              "$ref": "#/components/schemas/BadCaseView"
            },
            "type": "array",
            "title": "Bad Cases"
          },
          "reference_answer": {
            "type": "string",
            "title": "Reference Answer"
          },
          "memory_materials": {
            "items": {
              "$ref": "#/components/schemas/MemoryMaterialView"
            },
            "type": "array",
            "title": "Memory Materials"
          },
          "criteria": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/CriterionView"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Criteria"
          },
          "status": {
            "$ref": "#/components/schemas/QuestionStatus"
          },
          "next_action": {
            "$ref": "#/components/schemas/NextAction"
          },
          "content_revision": {
            "type": "integer",
            "title": "Content Revision"
          },
          "active_operation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Operation Id"
          },
          "last_error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GenerationErrorView"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          },
          "published_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "scene_id",
          "scene_name",
          "client_case_id",
          "title",
          "task_prompt",
          "reference_examples",
          "bad_cases",
          "reference_answer",
          "memory_materials",
          "criteria",
          "status",
          "next_action",
          "content_revision",
          "active_operation_id",
          "last_error",
          "created_at",
          "updated_at",
          "published_at"
        ],
        "title": "QuestionDetailResponse"
      },
      "QuestionLibraryResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/QuestionListItem"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "QuestionLibraryResponse"
      },
...
      "QuestionSaveRegenerateRequest": {
        "properties": {
          "command_id": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Command Id"
          },
          "content_revision": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Content Revision"
          },
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "task_prompt": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100000,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Prompt"
          },
          "reference_examples": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ReferenceExampleIn"
                },
                "type": "array",
                "maxItems": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference Examples"
          },
          "bad_cases": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/BadCaseIn"
                },
                "type": "array",
                "maxItems": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "Bad Cases"
          },
          "reference_answer": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200000,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference Answer"
          },
          "memory_materials": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/MemoryMaterialIn"
                },
                "type": "array",
                "maxItems": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "Memory Materials"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "command_id",
          "content_revision"
        ],
        "title": "QuestionSaveRegenerateRequest"
      },
      "QuestionStatus": {
        "type": "string",
        "enum": [
          "generating",
          "pending_review",
          "generation_failed",
          "published"
        ],
        "title": "QuestionStatus"
      },
      "QuestionTitleRequest": {
        "properties": {
          "command_id": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Command Id"
          },
          "content_revision": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Content Revision"
          },
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Title"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "command_id",
          "content_revision",
          "title"
        ],
        "title": "QuestionTitleRequest"
      },
...
      "RegisterRequest": {
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 50,
            "minLength": 2,
            "title": "Username"
          },
          "email": {
            "anyOf": [
              {
                "type": "string",
                "format": "email"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 8,
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "RegisterRequest"
      },
...
      "User": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "email": {
            "anyOf": [
              {
                "type": "string",
                "format": "email"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "username",
          "created_at"
        ],
        "title": "User"
      },
      "UserResponse": {
        "properties": {
          "user": {
            "$ref": "#/components/schemas/User"
          }
        },
        "type": "object",
        "required": [
          "user"
        ],
        "title": "UserResponse"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}